Implemented KeepAlive and Subscribe methods in TrackingService
In `Program.cs`, client creation for `TrackingService` has been modified to use the service's fully qualified name. Added `SendMessage` and `KeepAlive` methods. `SendMessage` sends a `TrackingMessage` to the server and handles `RpcException`. `KeepAlive` sends a stream of `PulseMessage` to the server.

In `TrackingService.proto`, added `KeepAlive` and `Subscribe` methods to `TrackingService`. Also added `PulseMessage`, `MachineStatus` enum, `SubscribeMessage`, and `Nofication` message.

In `MessageTrackingService.cs`, implemented `KeepAlive` and `Subscribe` methods. `KeepAlive` logs each `PulseMessage` received from the client. `Subscribe` sends a stream of `Nofication` messages to the client. Added a check in `ReviceMessage` method to ensure `DateTime` field of `TrackingMessage` is not null, throwing an `RpcException` with `InvalidArgument` status if it is.